Break Collections listing page Units into their own component
