+++ app/javascript/components/collections/list/CollectionsSortedByUnit.js
@@ -1,12 +1,9 @@
-import React, { useState } from 'react';
-import CollectionCard from './CollectionCard';
+import React from 'react';
 import '../Collection.scss';
-import ButtonCollectionListShowAll from './ButtonCollectionListShowAll';
 import PropTypes from 'prop-types';
+import CollectionsListUnit from './Unit';
 
 const CollectionsSortedByUnit = ({ filteredResult, sortByAZ, maxItems }) => {
-  const [showAll, setShowAll] = useState(false);
-
   const groupByUnit = list => {
     const map = new Map();
     list.forEach(item => {
@@ -30,50 +27,15 @@ const CollectionsSortedByUnit = ({ filteredResult, sortByAZ, maxItems }) => {
     return groups;
   };
 
-  const handleShowAll = event => {
-    setShowAll(!showAll);
-  };
-
-  return groupByUnit(filteredResult).map((unitArr, index) => {
-    let unit = unitArr[0];
-    let collections = sortByAZ(unitArr[1]);
-
-    return (
-      <section key={unit} className="collection-unit-wrapper">
-        <h2 className="headline collection-list-unit-headline">{unit}</h2>
-
-        <div className="row">
-          {collections.slice(0, maxItems).map((col, index) => {
-            return (
-              <div className="col-sm-4" key={col.id}>
-                <CollectionCard attributes={col} showUnit={false} />
-              </div>
-            );
-          })}
-        </div>
-
-        <div className="collapse" id={'collapse' + index}>
-          <div className="row">
-            {collections.slice(maxItems, collections.length).map(col => {
-              return (
-                <div className="col-sm-4" key={col.id}>
-                  <CollectionCard attributes={col} showUnit={false} />
-                </div>
-              );
-            })}
-          </div>
-        </div>
-
-        <ButtonCollectionListShowAll
-          collectionsLength={collections.length}
-          maxItems={maxItems}
-          handleShowAll={handleShowAll}
-          index={index}
-          showAll={showAll}
-        />
-      </section>
-    );
-  });
+  return groupByUnit(filteredResult).map((unitArr, index) => (
+    <CollectionsListUnit
+      key={unitArr[0]}
+      unitArr={unitArr}
+      index={index}
+      sortByAZ={sortByAZ}
+      maxItems={maxItems}
+    />
+  ));
 };
 
 CollectionsSortedByUnit.propTypes = {
